LOTE_GD/requirements.txt
#Command:
#pip install -r requirements.txt
python-dotenv==0.19.2
mysql-connector-python==8.0.26
Unidecode==1.3.2
XlsxWriter==3.0.1